+++ CopilotKit/packages/runtime-client-gql/CHANGELOG.md
@@ -1,5 +1,55 @@
 # @copilotkit/runtime-client-gql
 
+## 1.6.0-next.7
+
+### Patch Changes
+
+- @copilotkit/shared@1.6.0-next.7
+
+## 1.6.0-next.6
+
+### Patch Changes
+
+- @copilotkit/shared@1.6.0-next.6
+
+## 1.6.0-next.5
+
+### Patch Changes
+
+- Updated dependencies [090203d]
+  - @copilotkit/shared@1.6.0-next.5
+
+## 1.6.0-next.4
+
+### Patch Changes
+
+- @copilotkit/shared@1.6.0-next.4
+
+## 1.6.0-next.3
+
+### Patch Changes
+
+- @copilotkit/shared@1.6.0-next.3
+
+## 1.6.0-next.2
+
+### Patch Changes
+
+- @copilotkit/shared@1.6.0-next.2
+
+## 1.6.0-next.1
+
+### Patch Changes
+
+- d833f4c: - fix: provide the ability to type interrupt event value
+  - @copilotkit/shared@1.6.0-next.1
+
+## 1.6.0-next.0
+
+### Patch Changes
+
+- @copilotkit/shared@1.6.0-next.0
+
 ## 1.5.20
 
 ### Patch Changes