+++ CopilotKit/packages/shared/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @copilotkit/shared
 
+## 1.6.0-next.7
+
+## 1.6.0-next.6
+
+## 1.6.0-next.5
+
+### Patch Changes
+
+- 090203d: - fix: use tryMap method to filter out possibly invalid items
+
+## 1.6.0-next.4
+
+## 1.6.0-next.3
+
+## 1.6.0-next.2
+
+## 1.6.0-next.1
+
+## 1.6.0-next.0
+
 ## 1.5.20
 
 ### Patch Changes